Update schedule_screen.dart
everything works smoothly. the app spawns and inits in a the forwardlest weekday. then the indexDay syncs with it. when moving around using the bottom bar arrows everything changes accordingly. if it is a weekend day the text will show a message saying it is either sat or sun and there is no school.